(refactor) splits new / change property form into 2 components
- Adds NewPropertyView.tsx to new Property folder
